prepare/requirements.txt
Faker==17.6.0
psycopg-binary==3.1.8
python-dotenv==1.0.0
